move customElements define calls to index.js and refactor
This moves the `customeElements.define` calls from each respective
element and into the index.js file. This is useful for if we wish to
load the components classes without defining the custom element.

In addition, this changes the pattern of how we register the custom
elements, to allow for HMR, as we did in
https://github.com/github/catalyst/pull/197.